Modificamos los atributos danio y vida para que sean double en lugar de int y tenga efecto el 5% extra de daño si estas parado en territorio enemigo

+++ src/main/java/fiuba/AlgoChess/Modelo/Ataque/AtaqueCerca.java
@@ -6,7 +6,7 @@ import fiuba.AlgoChess.Modelo.Tablero.Casillero.Casillero;
 public class AtaqueCerca extends Habilidad {
 	
 	
-	public AtaqueCerca(int poder) {
+	public AtaqueCerca(double poder) {
 		
 		super(poder);
 		this.alcance = new Alcance(1, 2);
